feat(gateway/main.go): add support for initializing rest service for user feat(conf/conf.go): add comments to clarify the purpose of each section in the code feat(etclient/etclient.go): add serviceRegisterEventObserver to notify listeners when service is registered feat(etclient/etclient.go): add RegisterService function to register current service into etcd directory feat(etclient/etclient.go): add AddServiceRegisterEventListener function to add listeners for service registration events feat(etclient/etclient.go): remove unused function withKeyNamespace feat(etclient/util.go): add GetRawKeysByPrefix function to get raw keys by prefix feat(etclient/util.go): add WatchKeysByPrefix function to watch keys by prefix feat(etclient/util.go): add GetKeyLastSegment function to get the last segment of a key feat(etclient/util.go): add ConvertStringToLeaseID function to convert string to LeaseID
feat(rest.go): add rest package with Service struct and related methods for managing service directory

This commit adds a new file `rest/rest.go` which contains the implementation of the `Service` struct and related methods for managing a service directory. The `Service` struct represents a service and has properties like `Name` and `Endpoints`. The `Setup` function is currently empty and can be used for any necessary setup in the future.

The `Service` struct has a method `updateServiceDirectoryThread` which continuously watches for changes in the service directory using the `etclient.WatchKeysByPrefix` function. It updates the `Endpoints` map based on the events received and logs the updates. The `UpdateServiceDirectory` method retrieves the current state of the service directory using `etclient.GetRawKeysByPrefix` and updates the `Endpoints` map accordingly. It also logs the updated directory.

The `NewService` function creates a new instance of the `Service` struct with the given name. It initializes the `Endpoints` map and calls `UpdateServiceDirectory` to populate it with the current state of the service directory. It then starts a goroutine to continuously update the `Endpoints` map using `updateServiceDirectoryThread`. It also registers an event listener using `etclient.AddServiceRegisterEventListener` to update the service directory whenever a service is registered.

These changes are made to provide functionality for managing a service directory and keeping it up to date.